added website logo and removed unnecesary unused code and updated the Ending page running
