+++ src/Components/EndPageContent.jsx
@@ -1,0 +1,8 @@
+/* eslint-disable react/prop-types */
+import React from 'react';
+
+export default function EndPageContentWin(){
+    return <div data-testid = 'end-page-content-win' className="flex justify-center my-20 text-3xl text-fuchsia-500 font-semibold" >
+        Feel the rush of winning the game. If you want to keep winning.
+    </div>
+}
