+++ src/main.jsx
@@ -4,8 +4,7 @@ import App from './App.jsx'
 import './index.css'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import LandingPage from './Components/LandingPage.jsx'
-import EndingPageLoss from './Components/EndingPageLoss.jsx'
-import EndingPageWin from './Components/EndingPageWin.jsx'
+import EndingPage from './Components/EndingPage.jsx'
 import Play from './Components/Play.jsx'
 
 const router = createBrowserRouter([
@@ -18,12 +17,8 @@ const router = createBrowserRouter([
         element : <LandingPage/>
       },
       {
-        path : "/CS455_GAME/GameLost",
-        element : <EndingPageLoss/>
-      },
-      {
-        path : "/CS455_GAME/GameWon",
-        element : <EndingPageWin/>
+        path : "/CS455_GAME/GameOver",
+        element : <EndingPage/>
       },
       {
         path : "/CS455_GAME/play",
